Fix negative values always treated as false
When casting a floating point value to an unsigned integer, negative
values are coerced to zero (rather than their absolute value).  Thus
when a conditional's test expression was negative, it was always treated
as "false" due to being cast to a uint1.  This fix compares the test
expression with (not equal) zero instead.